szombat

+++ hazifeladat/hazifeladat.DAL1/Models/SeasonalRules.cs
@@ -10,7 +10,7 @@ namespace hazifeladat.DAL1.Models
         public int Id { get; set; }
         public DateTime From { get; set; }
         public DateTime To { get; set; }
-        public decimal Multiplier { get; set; }
+        public float Multiplier { get; set; }
         public ICollection<PlaceTypes>? AppliesToType { get; set; }
         public string? Name { get; set; }
     }
